refactor(input): move component and story file from tailwind folder
